website/requirements.txt
beautifulsoup4==4.7.1
cachetools==3.1.1
certifi==2018.11.29
chardet==3.0.4
datefinder==0.7.0
defusedxml==0.5.0
Django==2.2.6
django-allauth==0.38.0
django-braces==1.13.0
django-datetimepicker==3.14
django-markdownx==2.0.28
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.0.4
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
html2markdown==0.1.7
html2text==2018.1.9
httplib2==0.12.0
idna==2.8
Markdown==3.1.1
mistletoe==0.7.1
mysqlclient==1.3.13
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.0.0
pickle-mixin==1.0.2
Pillow==5.4.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyJWT==1.7.1
python-dateutil==2.7.5
python-decouple==3.1
python-social-auth==0.2.21
python3-openid==3.1.0
pytz==2018.9
regex==2019.11.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==1.2.0
socialauth==0.2.0
soupsieve==1.8
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
virtualenv==16.2.0
